First cut at error handling.
Attempt to catch errors in `run_ping` and send them back to the parent
so we can produce error messages in a controlled way.

There is still a problem where this occasionally fails and hangs, so I
still have some work to do to handle some of the problems.  In that case
there are errors like this happen:

    19:54:07.013 [error] Task #PID<0.135.0> started from #PID<0.48.0> terminating
    Function: &Ping.ping_async/2
        Args: ["192.168.1.79", #PID<0.48.0>]
    ** (exit) an exception was raised:
        ** (UndefinedFunctionError) undefined function:
    ErlangError.normalize/2
            (elixir) ErlangError.normalize(:emfile, nil)
            09-ping.exs:46: Ping.run_ping/1
            09-ping.exs:31: Ping.ping_async/2
            (elixir) lib/task/supervised.ex:74: Task.Supervised.do_apply/2
            (stdlib) proc_lib.erl:237: :proc_lib.init_p_do_apply/3

    19:54:07.016 [error] 'File operation error: emfile. Target: /usr/local/Cellar/elixir/1.0.5/bin/../lib/elixir/ebin/Elixir.ErlangError.beam.  Function: get_file. Process: code_server.'